+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## 2022-01-23 - Edu
+### Added
+- **processes/run_processes** Add envp to run_processes
 ### Changed
 - rename piped_process as process & t_pp as t_p
 - create_forkedpipes renamed as **run_processes**